Experiment with aliases
Page rule at CloudFlare will redirect everything from cbea.ms to
chris.beams.io with paths intact. If the path is an alias, it will be
meta-refreshed to it's fully expanded path.

Example:

 http://cbea.ms/a/hello -> 302 -> http://chris.beams.io/a/hello
 http://chris.beams.io/a/hello -> meta-refresh -> http://chris.beams.io/writing/hello-world

It would of course be ideal to have only one 302 hop and avoid the
in-browser refresh, but this appears to be the simplest/best arrangement
possible using Jekyll.